+++ comics/comics/pondus.py
@@ -16,6 +16,8 @@ class Crawler(CrawlerBase):
     time_zone = 1
 
     def crawl(self, pub_date):
-        url = 'http://www.dagbladet.no/tegneserie/pondusarkiv/serve.php?%s' % (
+        page_url = 'http://www.dagbladet.no/tegneserie/pondus/?%s' % (
             self.date_to_epoch(pub_date),)
+        page = self.parse_page(page_url)
+        url = page.src('img#pondus-stripe')
         return CrawlerImage(url)