+++ docs/changes.rst
@@ -16,6 +16,7 @@ v1.1.5 (in development)
 - Update: ``duelinganalogs`` was broken due to feed changes.
 - Update: ``fagprat`` has a new schedule.
 - Update: ``fanboys`` was broken due to feed changes.
+- Update: ``heltnils`` has a new schedule.
 
 
 v1.1.4 (2012-04-07)