+++ docs/changes.rst
@@ -17,6 +17,7 @@ v1.1.5 (in development)
 - Update: ``fagprat`` has a new schedule.
 - Update: ``fanboys`` was broken due to feed changes.
 - Update: ``heltnils`` has a new schedule.
+- Update: ``playervsplayer`` was broken due to feed changes.
 
 
 v1.1.4 (2012-04-07)